# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoffeeAndStudy API",
    "description": "API REST para gestionar lugares de estudio, usuarios, reseñas y sesiones de estudio",
    "contact": {
      "name": "Ibone",
      "email": "ibone23@example.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/study-sessions/{id}": {
      "get": {
        "tags": [
          "Study Sessions"
        ],
        "summary": "Obtener sesión de estudio por id",
        "description": "Devuelve una sesión concreta a partir de su identificador",
        "operationId": "getStudySessionById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sesión encontrada correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudySession"
                }
              }
            }
          },
          "404": {
            "description": "Sesión no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudySession"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudySession"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Study Sessions"
        ],
        "summary": "Actualizar sesión de estudio",
        "description": "Actualiza una sesión existente a partir de su id",
        "operationId": "updateStudySession",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudySession"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sesión actualizada correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudySession"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudySession"
                }
              }
            }
          },
          "404": {
            "description": "Sesión no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudySession"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudySession"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Study Sessions"
        ],
        "summary": "Eliminar sesión de estudio",
        "description": "Elimina una sesión de estudio a partir de su id",
        "operationId": "deleteStudySession",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sesión eliminada correctamente"
          },
          "404": {
            "description": "Sesión no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/reviews/{id}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Obtener reseña por id",
        "description": "Devuelve una reseña concreta a partir de su identificador",
        "operationId": "getReviewById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reseña encontrada correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "404": {
            "description": "Reseña no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Reviews"
        ],
        "summary": "Actualizar reseña",
        "description": "Actualiza una reseña existente a partir de su id",
        "operationId": "updateReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Review"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Reseña actualizada correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "404": {
            "description": "Reseña no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Reviews"
        ],
        "summary": "Eliminar reseña",
        "description": "Elimina una reseña a partir de su id",
        "operationId": "deleteReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Reseña eliminada correctamente"
          },
          "404": {
            "description": "Reseña no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
    "/study-sessions": {
      "get": {
        "tags": [
          "Study Sessions"
        ],
        "summary": "Obtener todas las sesiones de estudio",
        "description": "Devuelve la lista de sesiones, con posibilidad de filtrar por placeId, userId y productivityLevel",
        "operationId": "getAllStudySessions",
        "parameters": [
          {
            "name": "placeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "productivityLevel",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de sesiones obtenida correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudySession"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parámetros de filtro inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudySession"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudySession"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Study Sessions"
        ],
        "summary": "Crear sesión de estudio",
        "description": "Crea una nueva sesión de estudio",
        "operationId": "createStudySession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudySession"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Sesión creada correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudySession"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudySession"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudySession"
                }
              }
            }
          }
        }
      }
    },
    "/reviews": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Obtener todas las reseñas",
        "description": "Devuelve la lista de reseñas, con posibilidad de filtrar por placeId, rating y studyFriendly",
        "operationId": "getAllReviews",
        "parameters": [
          {
            "name": "placeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "rating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "studyFriendly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de reseñas obtenida correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parámetros de filtro inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "Crear reseña",
        "description": "Crea una nueva reseña",
        "operationId": "createReview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Review"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Reseña creada correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string",
            "maxLength": 30,
            "minLength": 3
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "maxLength": 100,
            "minLength": 8
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "premium": {
            "type": "boolean"
          },
          "registrationDate": {
            "type": "string",
            "format": "date"
          },
          "reputation": {
            "type": "number",
            "format": "float"
          }
        },
        "required": [
          "age",
          "email",
          "password",
          "premium",
          "username"
        ]
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "maxLength": 300,
            "minLength": 0
          },
          "colorHex": {
            "type": "string",
            "minLength": 1,
            "pattern": "^#([A-Fa-f0-9]{6})$"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date"
          }
        },
        "required": [
          "active",
          "colorHex",
          "description",
          "name",
          "priority"
        ]
      },
      "Place": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "maximum": 90,
            "minimum": -90
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "maximum": 180,
            "minimum": -180
          },
          "city": {
            "type": "string"
          },
          "wifiAvailable": {
            "type": "boolean"
          },
          "averagePrice": {
            "type": "number",
            "format": "float"
          },
          "openingDate": {
            "type": "string",
            "format": "date"
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "description",
          "latitude",
          "longitude",
          "title"
        ]
      },
      "StudySession": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "sessionDate": {
            "type": "string",
            "format": "date"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "productivityLevel": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          },
          "consumedSomething": {
            "type": "boolean"
          },
          "totalSpent": {
            "type": "number",
            "format": "float"
          },
          "reservedTable": {
            "type": "boolean"
          },
          "notes": {
            "type": "string",
            "maxLength": 300,
            "minLength": 0
          },
          "place": {
            "$ref": "#/components/schemas/Place"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "consumedSomething",
          "durationMinutes",
          "productivityLevel",
          "reservedTable",
          "sessionDate"
        ]
      },
      "Review": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "rating": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          },
          "comment": {
            "type": "string",
            "maxLength": 300,
            "minLength": 0
          },
          "studyFriendly": {
            "type": "boolean"
          },
          "noiseLevel": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          },
          "socketAvailability": {
            "type": "boolean"
          },
          "reviewDate": {
            "type": "string",
            "format": "date"
          },
          "recommendedHours": {
            "type": "number",
            "format": "float"
          },
          "place": {
            "$ref": "#/components/schemas/Place"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "comment",
          "noiseLevel",
          "rating",
          "socketAvailability",
          "studyFriendly"
        ]
      },
...
    }
  }
}